add prey sdk source code

+++ Q3E/src/main/jni/doom3/neo/quake4/ai/AAS_routing.cpp
@@ -145,7 +145,7 @@ void idAASLocal::CalculateAreaTravelTimes(void) {
 		}
 	}
 
-	assert( ( (unsigned long) bytePtr - (unsigned long) areaTravelTimes ) <= numAreaTravelTimes * sizeof( unsigned short ) );
+	assert( ( (uintptr_t) bytePtr - (uintptr_t) areaTravelTimes ) <= numAreaTravelTimes * sizeof( unsigned short ) );
 }
 
 /*
